lots of update to mail
Change from using smtpmail to message-send-mail-with-sendmail for
sending mail.  this offloads to sending mail to an external function.
Disable smtpmail use-package definition.
Change to using sendmail package.  Using a personal script to talk to
smtp servers because msmtp doesn't work the way I need it to with
OAUTH2.
Update notmuch search queries.