Major overhaul of emacs configuration
This is a big change to go back to an org-babel based configuration, where the
config is tangled from the Org file to the emacs .el files.  The difference from
last time is that this tangles straight to init.el rather than doing an
org-babel-load-file.

This also purges all the lisp that was under modules/ since it is now in
emacs.org.

It seems startup time is a bit better and this will allow for byte-compiling.