Continue overhaul of emacs to use straight
This further completes the work to migrate to straight as the package manager
for this configuration.  Some of the changes are merely reorganizing the
configuration into a, maybe, more logical config.  Changes to various packages
are a mix of removing the old mechanisms to install (:ensure), removing :defer,
clean up what are built-in packages vs those to install or are features, or how
features are part of larger packages.